Fixed incorrect statements in double exponential kernal synapse and buffer write back implementation. Fixed possion distribution. The framework generate expected results on an possion-neuron-to-lif-neuron example.

+++ include/interfaces/function.hpp
@@ -1,8 +1,8 @@
 #ifndef FUNCTION_HPP
 #define FUNCTION_HPP
 #include <functional>
-using NeuronDynamicsModel = std::function<std::vector<double>(int neuron_id, double input, double* state, double t, double* parameters, double dt)>;
-using SynapseDynamicsModel = std::function<std::vector<double>(double input, double* state, double t, double* parameters, double dt)>;
+using NeuronDynamicsModel = std::function<std::vector<double>(int neuron_id, double input, double* state, int t, double* parameters, double dt)>;
+using SynapseDynamicsModel = std::function<std::vector<double>(double input, double* state, int t, double* parameters, double dt)>;
 
 typedef double (*SynapseKernel)(int t, double tau);
 #endif
